Fix some minor warnings

+++ src/centurion/events/display_event.hpp
@@ -54,7 +54,7 @@ class display_event final : public common_event<SDL_DisplayEvent>
    */
   void set_event_id(const display_event_id id) noexcept
   {
-    m_event.event = to_underlying(id);
+    m_event.event = static_cast<u8>(to_underlying(id));
   }
 
   /**
